chore: Update hoppscotch docker-compose-aio.yaml
Update the hoppscotch docker-compose-aio.yaml file to modify the ports and environment variables. The ports have been changed to '13000:3000', '13100:3100', and '13170:3170'. The environment variables have been moved to an env_file and the file '.env' is now used. Additionally, a new service 'postgres' has been added with its own configuration for a PostgreSQL database.